Task4

+++ Assets/Content/Features/AIModule/Scripts/Entity/EntityData.cs
@@ -4,8 +4,8 @@ namespace Content.Features.AIModule.Scripts.Entity {
     [Serializable]
     public class EntityData {
         public EntityType EntityType;
-        public float StartHealth;
-        public float Damage;
+        public int StartHealth;
+        public int Damage;
         public float AttackDistance;
         public float InteractDistance;
         public float Speed;
