+++ Assets/Content/Features/DamageablesModule/Scripts/IDamageable.cs
@@ -11,7 +11,7 @@ namespace Content.Features.DamageablesModule.Scripts {
 
         public event Action OnDamaged;
         public event Action OnKilled;
-        public void Damage(float damage);
-        public void SetHealth(float health);
+        public void Damage(int damage);
+        public void SetHealth(int health);
     }
 }
